requirements.txt
flask
transformers
google-generativeai
torch
gunicorn
werkzeug